Arrumei o shuffle
