+++ package.json
@@ -8,6 +8,7 @@
     "crypto-js": "^4.0.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-markdown": "^5.0.3",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
